Added Shuffle functionality
The shapes on the game field can now be shuffled, which happens either when there are no more possible moves or by presing the shuffle button (for debug purposes only). This is accomplished by randomizing the shapes array using the Fisher-Yates algorithm and then placing the shapes into their newly assigned positions on the field. After shuffling, the game checks wether the shuffle resulted in matches and removes them accordingly.