Added Backgrounds and Win Screen
The Background Image of the Game Field can now be cahnged in the MainMenu. The Image Sprite gets scaled to fit the whole screen but using a sprite mask, only the part behind the Game Field is visible during play and it is partially gryed out. When the player wins the level the sprite mask is removed and the sprite's color is set to white, so the player can see the whole image clearly.