Added basic Drink Timer and Reward for Drinking
A timer counts down (currently from 5 minutes) and when it reaches 0 a notifcation pops up, asking the user to drink. Pressing OK grants the user a "Bomb Powerup" which destroys all objects in a 3x3 grid around the next selected object. This currently destroys Bonus shapes without triggering their bonus, which needs to be adressed.