First Steps of Shape Customization implemented
"Customization" Scene added. Here custom shapes can be cropped out of pictures. This is done by using a masking sprite and copying all pixels that lie inside of the mask from the source image to a new texture. Source and Mask Sprite have to have the Pivot in the bottom left and "Read/Write" on.