Added Customization UI to game
Customization UI is a seperate scene that can be loaded in the settings menu. There custom shapes can be cut out of Images that one can select from the Gallery and then saved locally. When the User then goes back to the Settings menu the saved images are loaded onto the custom shapes. if the custom shapes toggle is active, the shapesManager uses the custom shapes during the game.

Bug: Resizing the Background Sprites doesn't work on certain images.